clippy: fix Week 22 debt — safety docs, unused vars, auto-fixes
Added # Safety docs to all unsafe stub functions in cuda_stub.rs. Fixed unused imports and variables. Ran clippy --fix for automated fixes across examples and tests. Remaining: 136 warnings (down from 193).

+++ pesti-runner/examples/test_numerical_conformance.rs
@@ -44,7 +44,7 @@ async fn main() -> pesti_runner::error::Result<()> {
     println!("Time: {} ms", result.eval_ms);
     println!(
         "Throughput: {:.1} tok/s",
-        result.generated_tokens as f64 / (result.eval_ms as f64 / 1000.0)
+        result.generated_tokens as f64 / (result.eval_ms / 1000.0)
     );
 
     // Decode tokens to text
